- patterns: migrated to x64

+++ trunk/tools/patterns/jpegload.cpp
@@ -33,9 +33,9 @@ int JpegLoad(
     char *filename,
     void(*AddScanlineData) (unsigned char *buffer, int stride, void * Param),
     unsigned char **DecodeBuffer,
-    int *DecodeBufferSize,
-    int *DecodeWidth,
-    int *DecodeHegith,
+    long *DecodeBufferSize,
+    long *DecodeWidth,
+    long *DecodeHegith,
     void * Param
 )
 {
